+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Changed
 
 - Improve documentation.
+- Improve output of srcode ls.
 
 ## [0.2.0] - 2021-01-17